pyproject.toml
[tool.poetry]
name = "cardboard"
version = "0.1.0"
description = "A dashboard for managing puzzles and tracking status during a puzzle hunt."
authors = ["Cardinality Puzzlers <cardinalitypuzzles@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
python-dotenv = "^0.19.2"
django-heroku = "^0.3.1"
google-api-python-client = "^2.33.0"
social-auth-app-django = "^5.0.0"
django-taggit = "^2.0.0"
url-normalize = "^1.4.3"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
celery = "^5.2.3"
redis = "^4.1.0"
"discord.py" = "^1.7.3"
django-celery-beat = "^2.4.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pre-commit = "*"
black = "^22.3.0"
coverage = "*"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
